fix_problem
